+++ frontend/package.json
@@ -29,7 +29,6 @@
 		"data-forge": "^1.10.2",
 		"dom-to-image-more": "^3.4.5",
 		"env-cmd": "^10.1.0",
-		"html2canvas": "^1.4.1",
 		"jotai": "^2.10.0",
 		"jotai-location": "^0.5.5",
 		"just-debounce-it": "^3.2.0",
@@ -48,7 +47,6 @@
 		"react-table": "^7.8.0",
 		"react-vega": "^7.6.0",
 		"seamless-scroll-polyfill": "^2.3.4",
-		"use-react-screenshot": "^4.0.0",
 		"vega": "^5.30.0",
 		"vega-lite": "^5.21.0",
 		"vite": "^5.4.7",
