+++ frontend/src/cards/ui/CardOptionsMenu.tsx
@@ -36,10 +36,10 @@ export default function CardOptionsMenu(props: CardOptionsMenuProps) {
 
   return (
     <div
-      className='mb-0 mr-0 flex flex-row-reverse pr-0 sm:mt-1 sm:pr-5 md:mr-1 hide-on-screenshot'
+      className='mb-0 mr-0 flex flex-row-reverse pr-0 sm:mt-1 sm:pr-5 md:mr-1'
       id={'card-options-menu'}
     >
-      <Tooltip title='Card export options'>
+      <Tooltip title='Card export options' className='hide-on-screenshot'>
         <IconButton onClick={shareMenu.open}>
           <MoreHorizIcon />
         </IconButton>
