+++ frontend/src/cards/ui/SourcesHelpers.tsx
@@ -1,5 +1,4 @@
 import type { MapOfDatasetMetadata } from '../../data/utils/DatasetTypes'
-import { HET_URL } from '../../utils/internalRoutes'
 import { dataSourceMetadataMap } from '../../data/config/MetadataMap'
 import type { MetricQueryResponse } from '../../data/query/MetricQuery'
 import type {
@@ -10,7 +9,6 @@ import type {
 export function currentYear(): number {
   return new Date().getFullYear()
 }
-export const CITATION_APA = `Health Equity Tracker. (${currentYear()}). Satcher Health Leadership Institute. Morehouse School of Medicine. ${HET_URL}.`
 
 export function insertPunctuation(idx: number, numSources: number) {
   let punctuation = ''
