+++ frontend/src/pages/WhatIsHealthEquity/FaqTab.tsx
@@ -1,8 +1,8 @@
 import { ABOUT_US_PAGE_LINK } from '../../utils/internalRoutes'
 import { Helmet } from 'react-helmet-async'
 import MoreHorizIcon from '@mui/icons-material/MoreHoriz'
-import { CITATION_APA } from '../../cards/ui/SourcesHelpers'
 import { selectFAQs } from './FaqData'
+import { CITATION_APA } from '../Methodology/methodologyComponents/MethodologyPage'
 
 function FaqTab() {
   return (
