+++ frontend/src/vite-env.d.ts
@@ -1,4 +1,1 @@
-// eslint-disable-next-line @typescript-eslint/triple-slash-reference
 /// <reference types="vite/client" />
-
-declare module 'use-react-screenshot'
